+++ lib/src/subtypes/guilds/guild_type.dart
@@ -1,3 +1,5 @@
+import 'package:randpg/entities/emblems.dart';
+
 import '../../generators/base/generator.dart';
 
 /// A class that represents a guild type
@@ -16,6 +18,9 @@ abstract class GuildType {
 
   /// Returns a generator for the history of the guild
   IGenerator<String> getHistoryGenerator(String guildName);
+
+  /// Returns the emblem type for this guild type
+  EmblemType getEmblemType();
 
   /// Returns a generator for a guild motto
   IGenerator<String> getMottoGenerator();